:sparkles: 에러 발생시 alert 구현
