+++ Diary/Sources/Extensions/UITextField+.swift
@@ -5,4 +5,19 @@
 //  Created by 이태영 on 2023/01/05.
 //
 
-import Foundation
+import UIKit
+
+extension UITextField {
+    convenience init(font: UIFont?, placeholder: String? = nil) {
+        self.init()
+        
+        self.font = font
+        self.placeholder = placeholder
+        self.adjustsFontForContentSizeCategory = true
+        self.returnKeyType = .done
+    }
+    
+    var filteredText: String {
+        return self.text ?? ""
+    }
+}
